+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Fixed a bug where clicking a tag in `sl-select` wouldn't toggle the menu
 - Fixed a bug where options where `sl-select` options weren't always visible or scrollable
+- Fixed a bug where setting `null` on `sl-input`, `sl-textarea`, or `sl-select` would throw an error
 
 ## 2.0.0-beta.16